tuesday-service-server-backup/requirements.txt
speedtest-cli==2.0.2
requests==2.19.1
mysql-connector==2.1.6
bitmath==1.3.1.2
youtube_dl==2018.10.05
paramiko==2.4.1